Lyon/requirements.txt
ppretty
requests
django-filter
Django>=2.2.6
djangorestframework
gunicorn
mysqlclient